backend/django-backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
Pillow==10.0.0
qrcode==7.3.1
django-cors-headers==3.14.0
firebase-admin==6.0.0
celery==5.3.0
django-i18nfield==3.0.0
psycopg2-binary==2.9.6
boto3==1.26.0